Fix 2 flaky tests:
- TestServerWebApp.getHomeDir
- TestViewfsWithNfs3.testNfsRenameSingleNN

Fix 1:
--------------------
Clear the system properties at the end of TestServerWebApp.getHomeDir to
prevent failure when running it twice.

Fix 2:
--------------------
Rename a file back to the original to prevent failure when running
TestViewfsWithNfs3.testNfsRenameSingleNN twice. Without the fix, the
second run of the test would fail as the original file can no longer be
accessed once it gets renamed in the first run.